Moved from php to html, eliminated old php scripts, added examples, tweaked css, and more.
Previously all pages were written with a .php extension, though this
was not a requirement. These have been renamed to their .html
equivalent, and all links updated.
The old PHP files in the scripts directory have been removed.
(Unnecessary.)
Tweaked CSS and added support for input elements (forms).
Added a TextEntry example.
Added a "content" example using the Bible as an example. This content
example uses a bottom-anchored black bar with in-built select lists to
select the bible, book, and chapter, whilst the content is in a typical
black-on-white frame.
Added caching via cache.manifest; right now the directive is to load
all content from the network. This will change after development nears
completion.
Updated iScroll 4 to handle SELECTs, INPUTs, and TEXTAREAs.
Also added a .htaccess file to control output of the manifest.